Updated notebooks

+++ src/model.py
@@ -81,16 +81,27 @@ class MLTModel(object):
         print("n_singletons: " + str(len(self.singletons)))
 
     def construct_network(self):
-        self.word_ids = tf.placeholder(tf.int32, [None, None], name="word_ids")
-        self.char_ids = tf.placeholder(tf.int32, [None, None, None], name="char_ids")
-        self.sentence_lengths = tf.placeholder(tf.int32, [None], name="sentence_lengths")
-        self.word_lengths = tf.placeholder(tf.int32, [None, None], name="word_lengths")
+        # self.word_ids = tf.placeholder(tf.int32, [None, None], name="word_ids")
+        # self.char_ids = tf.placeholder(tf.int32, [None, None, None], name="char_ids")
+        # self.sentence_lengths = tf.placeholder(tf.int32, [None], name="sentence_lengths")
+        # self.word_lengths = tf.placeholder(tf.int32, [None, None], name="word_lengths")
+        # self.sentence_labels = tf.placeholder(tf.float32, [None,], name="sentence_labels")
+        # self.word_labels = tf.placeholder(tf.float32, [None,None], name="word_labels")
+        # self.word_objective_weights = tf.placeholder(tf.float32, [None,None], name="word_objective_weights")
+        # self.sentence_objective_weights = tf.placeholder(tf.float32, [None], name="sentence_objective_weights")
+        # self.learningrate = tf.placeholder(tf.float32, name="learningrate")
+        # self.is_training = tf.placeholder(tf.int32, name="is_training")
+
+        self.word_ids = tf.placeholder(tf.float32, [None, None], name="word_ids")
+        self.char_ids = tf.placeholder(tf.float32, [None, None, None], name="char_ids")
+        self.sentence_lengths = tf.placeholder(tf.float32, [None], name="sentence_lengths")
+        self.word_lengths = tf.placeholder(tf.float32, [None, None], name="word_lengths")
         self.sentence_labels = tf.placeholder(tf.float32, [None,], name="sentence_labels")
         self.word_labels = tf.placeholder(tf.float32, [None,None], name="word_labels")
         self.word_objective_weights = tf.placeholder(tf.float32, [None,None], name="word_objective_weights")
         self.sentence_objective_weights = tf.placeholder(tf.float32, [None], name="sentence_objective_weights")
         self.learningrate = tf.placeholder(tf.float32, name="learningrate")
-        self.is_training = tf.placeholder(tf.int32, name="is_training")
+        self.is_training = tf.placeholder(tf.float32, name="is_training")
 
         self.loss = 0.0
         input_tensor = None
